+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All Notable changes to `Period Visualizer` will be documented in this file
 - Adding `Dataset::appendAll`
 - Adding `ConsoleConfig::leftMargin`
 - Adding `ConsoleConfig::withLeftMargin`
+- Renamed `ConsoleGraph` to `GanttBar`
+- Renamed `ConsoleConfig` to `GanttBarConfig`
+- Removed `Pairs` interface
 
 ## 0.5.1 - 2019-07-28